+++ middleware/middleware.go
@@ -1,68 +1,23 @@
 package middleware
 
 import (
-	"context"
-	"net/http"
-	"strings"
+	"errors"
 
-	firebase "firebase.google.com/go"
-	"firebase.google.com/go/auth"
 	"github.com/gin-gonic/gin"
-	"google.golang.org/api/option"
+	"github.com/hackhack-Geek-vol6/backend/server"
 )
 
 const (
-	authorizationHeader = "Authorization"
-	authorizationType   = "bearer"
-	valName             = "token"
+	authorizationHeaderKey  = "authorization"
+	authorizationTypeBearer = "bearer"
 )
 
-type FirebaseAuthMiddleware struct {
-	client       *auth.Client
-	unAuthorized func(ctx *gin.Context)
-}
-
-func New(credFileName string, unAuthorized func(ctx *gin.Context)) (*FirebaseAuthMiddleware, error) {
-	opt := option.WithCredentialsFile(credFileName)
-	app, err := firebase.NewApp(context.Background(), nil, opt)
-	if err != nil {
-		return nil, err
-	}
-	auth, err := app.Auth(context.Background())
-	if err != nil {
-		return nil, err
-	}
-	return &FirebaseAuthMiddleware{
-		client:       auth,
-		unAuthorized: unAuthorized,
-	}, nil
-}
-func (firebaseAuthMiddleware *FirebaseAuthMiddleware) MiddlewareFunc() gin.HandlerFunc {
+func AuthMiddleware(server *server.Server) gin.HandlerFunc {
 	return func(ctx *gin.Context) {
-		authHeader := ctx.Request.Header.Get(authorizationHeader)
-		token := strings.Replace(authHeader, authorizationType, "", 1)
-		idToken, err := firebaseAuthMiddleware.client.VerifyIDToken(context.Background(), token)
-		if err != nil {
-			if firebaseAuthMiddleware.unAuthorized != nil {
-				firebaseAuthMiddleware.unAuthorized(ctx)
-				ctx.Abort()
-				return
-			}
-			ctx.JSON(http.StatusUnauthorized, gin.H{
-				"status":  http.StatusUnauthorized,
-				"message": http.StatusText(http.StatusUnauthorized),
-			})
-			return
-		}
-		ctx.Set(valName, idToken)
-		ctx.Next()
-	}
-}
+		authorizationHeader := ctx.GetHeader(authorizationHeaderKey)
 
-func ExtractClaims(c *gin.Context) *auth.Token {
-	idToken, ok := c.Get(valName)
-	if !ok {
-		return new(auth.Token)
+		if len(authorizationHeader) == 0 {
+			err := errors.New("authorization header is not provided")
+		}
 	}
-	return idToken.(*auth.Token)
 }
