+++ server/server.go
@@ -4,7 +4,6 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/hackhack-Geek-vol6/backend/controller"
 	db "github.com/hackhack-Geek-vol6/backend/db/sqlc"
-	"github.com/hackhack-Geek-vol6/backend/middleware"
 	"github.com/hackhack-Geek-vol6/backend/util"
 )
 
@@ -29,29 +28,6 @@ func NewServer(config util.EnvConfig, store db.Store) (*Server, error) {
 	}
 
 	return server, nil
-}
-
-// ルーティングをセットアップする
-func (server *Server) setupRouter() {
-	router := gin.New()
-
-	router.GET("/v1/ping", server.controller.Ping)
-
-	server.router = router
-}
-
-// 認証ミドルウェアの必要なルーティング
-func (server *Server) authRouter() error {
-	middleware, err := middleware.New("./credentials.json", nil)
-	if err != nil {
-		return err
-	}
-
-	auth := server.router.Group("/auth")
-	auth.Use(middleware.MiddlewareFunc())
-	auth.GET("/v1/pong", server.controller.Pong)
-
-	return nil
 }
 
 // サーバを開始する
